generate_html_files()を修正
