--- tests/generate/mod.rs
@@ -1,6 +1,0 @@
-use rustatic::generate;
-
-#[test]
-fn htmls_generate_test() {
-    generate::generate_html_files();
-}
